+++ test/bract/core/inducer_test.clj
@@ -47,13 +47,13 @@
         (Echo/setVerbose verbosity?)))))
 
 
-(deftest test-run-inducers
+(deftest test-run-config-inducers
   (let [verbosity? (Echo/isVerbose)
         context {:bract.core/verbose? true
                  :bract.core/config {"bract.core.inducers" ['bract.core.inducer/set-verbosity]}}]
     (try
       (Echo/setVerbose false)
-      (inducer/run-inducers context)
+      (inducer/run-config-inducers context)
       (is (true? (Echo/isVerbose)) "Verbosity should be enabled after configuring it as true")
       (finally
         (Echo/setVerbose verbosity?)))))